+++ cpskin/core/tests/test_banners.py
@@ -1,13 +1,24 @@
 # -*- coding: utf-8 -*-
+
+from Products.Five.browser import BrowserView
 from cpskin.core.testing import CPSKIN_CORE_INTEGRATION_TESTING
 from plone import api
-from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
-from Products.Five.browser import BrowserView
+from plone.app.testing import setRoles
+from plone.namedfile.file import NamedBlobImage
 from zope.component import queryMultiAdapter
 from zope.viewlet.interfaces import IViewletManager
 
+import os
 import unittest
+
+
+def dummy_image():
+    filename = os.path.join(os.path.dirname(__file__), u'banner.png')
+    return NamedBlobImage(
+        data=open(filename, 'r').read(),
+        filename=filename
+    )
 
 
 class TestBanners(unittest.TestCase):
@@ -30,18 +41,19 @@ class TestBanners(unittest.TestCase):
 
     def test_banner_without_folder(self):
         url = self.my_viewlet.getImageBannerUrl()
-        self.assertEqual(url, u'http://nohost/plone/banner.jpg')
+        self.assertTrue(url.startswith('http://nohost/plone/banner.jpg'))
 
     def test_banner_with_folder_without_image(self):
         api.content.create(
             container=self.portal, type='Folder', id='banner')
         url = self.my_viewlet.getImageBannerUrl()
-        self.assertEqual(url, u'http://nohost/plone/banner.jpg')
+        self.assertTrue(url.startswith('http://nohost/plone/banner.jpg'))
 
     def test_banner_with_folder_with_image(self):
         folder = api.content.create(
             container=self.portal, type='Folder', id='banner')
-        api.content.create(
-            container=folder, type='Image', id='image1.png')
+        image = api.content.create(
+            container=folder, type='Image', id='img1.png')
+        image.image = dummy_image()
         url = self.my_viewlet.getImageBannerUrl()
-        self.assertEqual(url, u'http://nohost/plone/banner/image1.png')
+        self.assertTrue(url.startswith('http://nohost/plone/banner/img1.png'))